requirements.txt
fastapi
uvicorn[standard]
beanie
motor
python-dotenv
pydantic[email]